requirements.txt
# Web Framework & Server
fastapi==0.135.1
python-multipart==0.0.22  # Required for UploadFile/File in FastAPI
uvicorn==0.41.0           # Required to actually run your FastAPI app

# AI & Machine Learning
torch==2.10.0
torchaudio==2.10.0
openai-whisper==20250625
transformers==4.57.6
indictranstoolkit==1.1.1
sentencepiece==0.2.1      # Often required under-the-hood by IndicTrans2
accelerate==1.12.0        # Helps Transformers run models efficiently

# Utilities & APIs
python-dotenv==1.2.2
groq==1.0.0